Title: 합과 평균의 차, Time: 42ms, Memory: 16MB, Status: Passed - Codetree

+++ 250820/합과 평균의 차/sub-of-average-and-sum.py
@@ -1,6 +1,6 @@
 a, b, c = map(int, input().split())
 sum = a + b + c
-avg = sum / 3
+avg = int(sum / 3)
 print(sum)
 print(avg)
 print(sum - avg)
